% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_fct.R
\name{.apply_fct_lookup}
\alias{.apply_fct_lookup}
\title{Apply a lookup table to a character vector}
\usage{
.apply_fct_lookup(.data, .lookup = NULL)
}
\arguments{
\item{.data}{(\code{character} or coercible to \code{character}) A vector to harmonize
to the specified factor.}

\item{.lookup}{(named \code{character}) A vector of replacement values. The names
are the values in \code{.data} and the values are the target values.}
}
\value{
A character vector with values replaced according to the lookup
table.
}
\description{
Apply a lookup table to a character vector
}
\keyword{internal}
