% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-conditions.R
\name{.check_args_spec}
\alias{.check_args_spec}
\title{Check that all args are hrmn_spec objects}
\usage{
.check_args_spec(..., call = rlang::caller_env())
}
\arguments{
\item{...}{Arguments to check.}

\item{call}{\code{(environment)} The execution environment to mention as the
source of error messages.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Check that all args are hrmn_spec objects
}
\keyword{internal}
