% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_df.R
\name{.harmonize_col}
\alias{.harmonize_col}
\title{Harmonize a specific column within a data.frame}
\usage{
.harmonize_col(.data, .spec, .col_name, .call = rlang::caller_env())
}
\arguments{
\item{.data}{(\code{data.frame}) A data frame to harmonize.}

\item{.spec}{(\code{hrmn_spec_df}) A data frame harmonization specification.}

\item{.col_name}{(\verb{length-1 character}) The name of the column to harmonize
within \code{.data}.}

\item{.call}{(\code{environment}) The execution environment to mention as the
source of error messages.}
}
\value{
The input \code{.data} harmonized to a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Harmonize a specific column within a data.frame
}
\keyword{internal}
