% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_df.R
\name{.harmonize_col_names}
\alias{.harmonize_col_names}
\title{Harmonize data frame column names against a specification}
\usage{
.harmonize_col_names(
  .data,
  .spec,
  .unspecified_columns,
  .call = rlang::caller_env()
)
}
\arguments{
\item{.data}{(\code{data.frame}) A data frame to harmonize.}

\item{.spec}{(\code{hrmn_spec_df}) A data frame harmonization specification.}

\item{.unspecified_columns}{(\code{"error"}, \code{"drop"}, or \code{"keep"}) How to handle
columns in \code{.data} that are not present in \code{.spec}.}

\item{.call}{(\code{environment}) The execution environment to mention as the
source of error messages.}
}
\value{
A \code{data.frame}, possibly with columns removed.
}
\description{
This will almost definitely migrate to stbl in the future.
}
\keyword{internal}
