% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_df.R
\name{.harmonize_df_data}
\alias{.harmonize_df_data}
\title{Harmonize a data frame based on data inputs}
\usage{
.harmonize_df_data(
  .data,
  ...,
  .spec,
  .unspecified_columns,
  .call = rlang::caller_env()
)
}
\arguments{
\item{.data}{(\code{data.frame}) A data frame to harmonize.}

\item{...}{These dots are for future extensions and must be empty.}

\item{.spec}{(\code{hrmn_spec_df}) A data frame harmonization specification.}

\item{.unspecified_columns}{(\code{"error"}, \code{"drop"}, or \code{"keep"}) How to handle
columns in \code{.data} that are not present in \code{.spec}.}

\item{.call}{(\code{environment}) The execution environment to mention as the
source of error messages.}
}
\value{
A \code{data.frame} with custom harmonizations and column name
reconciliation applied.
}
\description{
Harmonize a data frame based on data inputs
}
\keyword{internal}
