% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_df.R
\name{.harmonize_df_dots}
\alias{.harmonize_df_dots}
\title{Harmonize a data frame based on custom harmonization calls}
\usage{
.harmonize_df_dots(.data, ..., .call = rlang::caller_env())
}
\arguments{
\item{.data}{(\code{data.frame}) A data frame to harmonize.}

\item{...}{These dots are for future extensions and must be empty.}

\item{.call}{(\code{environment}) The execution environment to mention as the
source of error messages.}
}
\value{
A \code{data.frame} with custom harmonizations applied.
}
\description{
Harmonize a data frame based on custom harmonization calls
}
\keyword{internal}
