% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_df.R
\name{.harmonize_df_spec}
\alias{.harmonize_df_spec}
\title{Harmonize a data frame based on a harmonization specification}
\usage{
.harmonize_df_spec(.data, .spec, .call = rlang::caller_env())
}
\arguments{
\item{.data}{(\code{data.frame}) A data frame to harmonize.}

\item{.spec}{(\code{hrmn_spec_df}) A data frame harmonization specification.}

\item{.call}{(\code{environment}) The execution environment to mention as the
source of error messages.}
}
\value{
The input \code{.data} harmonized to a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Harmonize a data frame based on a harmonization specification
}
\keyword{internal}
