% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_df.R
\name{.stop_col_has_unknown_spec}
\alias{.stop_col_has_unknown_spec}
\title{Stop if a column has an unknown specification class}
\usage{
.stop_col_has_unknown_spec(.col_name, class, .call = rlang::caller_env())
}
\arguments{
\item{.col_name}{(\verb{length-1 character}) The name of the column.}

\item{class}{(\code{character}) The classes of the column specification.}

\item{.call}{(\code{environment}) The execution environment to mention as the
source of error messages.}
}
\description{
Stop if a column has an unknown specification class
}
\keyword{internal}
