% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-conditions.R
\name{.stop_if_args_unnamed}
\alias{.stop_if_args_unnamed}
\title{Check that all specified args are named}
\usage{
.stop_if_args_unnamed(..., .call = rlang::caller_env())
}
\arguments{
\item{...}{Arguments to check.}

\item{.call}{(\code{environment}) The execution environment to mention as the
source of error messages.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Check that all specified args are named
}
\keyword{internal}
