% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_df.R
\name{.stop_if_missing_col_names}
\alias{.stop_if_missing_col_names}
\title{Stop if data frame is missing columns from specification}
\usage{
.stop_if_missing_col_names(.data, .spec, .call = rlang::caller_env())
}
\arguments{
\item{.data}{(\code{data.frame}) A data frame to harmonize.}

\item{.spec}{(\code{hrmn_spec_df}) A data frame harmonization specification.}

\item{.call}{(\code{environment}) The execution environment to mention as the
source of error messages.}
}
\value{
\code{NULL} (invisibly)
}
\description{
Stop if data frame is missing columns from specification
}
\keyword{internal}
