% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_hrmn_spec.R
\name{.to_hrmn_spec}
\alias{.to_hrmn_spec}
\title{Validate a harmonization specification}
\usage{
.to_hrmn_spec(.spec, .subclass = NULL, .call = rlang::caller_env())
}
\arguments{
\item{.spec}{(\code{NULL} or \code{hrmn_spec}) An object to validate as a harmonization
specification object.}

\item{.subclass}{(\code{NULL} or \verb{length-1 character}) If provided, check that
\code{.spec} inherits the given subclass. For example, to confirm that \code{.spec}
has class \verb{"hrmn_spec_fct}, use \code{.subclass = "fct"}.}
}
\value{
The validated \code{.spec} object, or an informative error.
}
\description{
Validate a harmonization specification
}
\keyword{internal}
