% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_df.R
\name{harmonize_df}
\alias{harmonize_df}
\title{Harmonize a data frame}
\usage{
harmonize_df(
  .data,
  ...,
  .spec = NULL,
  .unspecified_columns = c("error", "drop", "keep")
)
}
\arguments{
\item{.data}{(\code{data.frame}) A data frame to harmonize.}

\item{...}{These dots are for future extensions and must be empty.}

\item{.spec}{(\code{hrmn_spec_df}) A data frame harmonization specification.}

\item{.unspecified_columns}{(\code{"error"}, \code{"drop"}, or \code{"keep"}) How to handle
columns in \code{.data} that are not present in \code{.spec}.}
}
\value{
The input \code{.data} harmonized to a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Harmonize a data frame
}
\examples{
df <- data.frame(
  size = c("Small", "Medium", "S", "M", "Large", "Lrg", "Sm"),
  id = 1:7
)

# This spec will coerce values to NA if they are not "Small", "Medium",
# or "Large".
spec <- specify_df(
  size = specify_fct(levels = c("Small", "Medium", "Large"))
)

# We can provide harmonization rules to the data before the spec is applied.
# Here, we harmonize the input factor to convert "S", "M", "Sm", and "Lrg" to
# valid values.
harmonize_df(
  df,
  size = harmonize_fct(
    size,
    .lookup = c("S" = "Small", "M" = "Medium", "Sm" = "Small", "Lrg" = "Large")
  ),
  .spec = spec,
  .unspecified_columns = "keep"
)
}
\seealso{
Other harmonization functions: 
\code{\link{harmonize_fct}()}
}
\concept{harmonization functions}
