% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_fct.R
\name{harmonize_fct}
\alias{harmonize_fct}
\title{Harmonize a factor}
\usage{
harmonize_fct(.data, ..., .spec = NULL, .lookup = NULL)
}
\arguments{
\item{.data}{(\code{character} or coercible to \code{character}) A vector to harmonize
to the specified factor.}

\item{...}{These dots are for future extensions and must be empty.}

\item{.spec}{(\code{hrmn_spec_fct}) A harmonization specification from
\code{\link[=specify_fct]{specify_fct()}}.}

\item{.lookup}{(named \code{character}) A vector of replacement values. The names
are the values in \code{.data} and the values are the target values.}
}
\value{
A harmonized \code{\link[=factor]{factor()}}.
}
\description{
Harmonize a factor
}
\examples{
# Without a spec, harmonize_fct() acts like [base::factor()].
harmonize_fct(c("a", "b", "c"))

# Basic harmonization, dropping levels not in the spec
spec <- specify_fct(levels = c("a", "b"))
harmonize_fct(c("a", "b", "c"), .spec = spec)

# Using a lookup table to recode values
spec2 <- specify_fct(levels = c("fruit", "citrus"))
lookup <- c(apple = "fruit", banana = "fruit", orange = "citrus")
harmonize_fct(
  c("apple", "banana", "orange"),
  .spec = spec2,
  .lookup = lookup
)
}
