% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrmn-package.R
\docType{package}
\name{hrmn-package}
\alias{hrmn}
\alias{hrmn-package}
\title{hrmn: Harmonize Datasets}
\description{
A common early step during data analysis is data harmonization: converting disparate datasets into a unified, consistent format, with consistent column names, classes, and values. The goal of 'hrmn' is to make this process as easy as it can be.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/hrmn}
  \item \url{https://jonthegeek.github.io/hrmn/}
  \item Report bugs at \url{https://github.com/jonthegeek/hrmn/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
