% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_df.R
\name{specify_df}
\alias{specify_df}
\title{Data frame specification}
\usage{
specify_df(...)
}
\arguments{
\item{...}{(\code{hrmn_spec}) Column specifications, given as named arguments.}
}
\value{
A \code{hrmn_spec_df} object that acts as a specification.
}
\description{
Create an object that specifies the desired format for a data frame.
This specification object does not contain any data itself, only the rules
for harmonization.
}
\examples{
specify_df(
  response = specify_fct(levels = c("Yes", "No", "Maybe")),
  outcome = specify_fct(levels = c("Positive", "Negative"))
)
}
\seealso{
Other specification functions: 
\code{\link{specify_fct}()}
}
\concept{specification functions}
