% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_fct.R
\name{specify_fct}
\alias{specify_fct}
\title{Specify a factor harmonization}
\usage{
specify_fct(levels = character())
}
\arguments{
\item{levels}{(\code{character}) The allowed values of the factor.}
}
\value{
A factor specification, an S7 object of class \code{hrmn::hrmn_fct}.
}
\description{
Create a \code{hrmn_fct} object that specifies the desired levels for a factor
variable. This 'specification' object does not contain any data itself, only
the rules for harmonization.
}
