% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared_params.R
\name{.shared_params}
\alias{.shared_params}
\title{Parameters used in multiple functions}
\arguments{
\item{levels}{(\code{character}) The allowed values of the factor.}
}
\description{
Reused parameter definitions are gathered here for easier editing.
}
\keyword{internal}
