% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_fct.R
\name{specify_fct}
\alias{specify_fct}
\title{Factor specification}
\usage{
specify_fct(levels = character())
}
\arguments{
\item{levels}{(\code{character}) The allowed values of the factor.}
}
\value{
A \code{hrmn_fct_spec} object that acts as a specification.
}
\description{
Create an object that specifies the desired levels for a factor variable.
This specification object does not contain any data itself, only the rules
for harmonization.
}
\examples{
specify_fct(levels = c("a", "b", "c"))
}
\seealso{
Other specification functions: 
\code{\link{specify_df}()}
}
\concept{specification functions}
